chore: merge database resource type declare
