+++ frontend/src/components/ProjectMember/ProjectMemberTable/types.ts
@@ -1,4 +1,4 @@
-import { DatabaseResource } from "@/components/Issue/form/SelectDatabaseResourceForm/common";
+import { DatabaseResource } from "@/types";
 import { User } from "@/types/proto/v1/auth_service";
 import { Binding } from "@/types/proto/v1/project_service";
