+++ frontend/src/types/member.ts
@@ -1,3 +1,9 @@
 export type MemberStatus = "INVITED" | "ACTIVE";
 
 export type RoleType = "OWNER" | "DBA" | "DEVELOPER";
+
+export interface DatabaseResource {
+  databaseName: string;
+  schema?: string;
+  table?: string;
+}
