+++ frontend/src/utils/issue/cel.ts
@@ -1,6 +1,6 @@
 import { cloneDeep, last } from "lodash-es";
 import { SimpleExpr, resolveCELExpr } from "@/plugins/cel";
-import { DatabaseResource } from "@/components/Issue/form/SelectDatabaseResourceForm/common";
+import { DatabaseResource } from "@/types";
 import { Expr } from "@/types/proto/google/api/expr/v1alpha1/syntax";
 import { celServiceClient } from "@/grpcweb";
